requirements.txt
fastapi
uvicorn
python-multipart
requests
langgraph
langchain-core
langchain-openai
pymupdf
pydantic
openai
streamlit
pandas
python-dotenv